Move service instance test from exp. to v3
Also added cleanup to service instance test suite
