+++ helpers/app_helpers/app_helpers.go
@@ -1,8 +1,9 @@
 package app_helpers
 
 import (
-	"strings"
 	"fmt"
+	"strings"
+
 	"github.com/cloudfoundry-incubator/cf-test-helpers/cf"
 	. "github.com/cloudfoundry/cf-acceptance-tests/cats_suite_helpers"
 	"github.com/cloudfoundry/cf-acceptance-tests/helpers/logs"
